Functionality updates
Major changes for task_list():
Added: Option to either view all, view in-progress or done + option to opt out.

Changed the task_update_status() func:
Instead of having to write the status "in-progress" or "done", it checks itself that the
status is, and switches it.

Added a load_tasks() func, for loading the json file to decrease amount of redundancy.

Trimmed tasks.json file.

Updated todo-list.